crud success

+++ BW_U_1/BW_U_1/Models/Products.cs
@@ -2,10 +2,12 @@
 {
     public class Products
     {
-        public int IdProd { get; set; } 
-        public string Name {  get; set; }
+        public int IdProd { get; set; }
+        public string Name { get; set; }
         public string DescriptionProd { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Category { get; set; }
+
+        
     }
 }
